+++ inc/cmb2/js/cmb2.js
@@ -15,7 +15,7 @@ window.CMB2 = (function(window, document, $, undefined){
 	var l10n = window.cmb2_l10;
 	var setTimeout = window.setTimeout;
 
-	// CMB functionality object
+	// CMB2 functionality object
 	var cmb = {
 		formfield       : '',
 		idNumber        : false,
@@ -23,6 +23,7 @@ window.CMB2 = (function(window, document, $, undefined){
 		repeatEls       : 'input:not([type="button"]),select,textarea,.cmb2-media-status',
 		styleBreakPoint : 450,
 		mediaHandlers   : {},
+		neweditor_id    : [],
 		defaults : {
 			time_picker  : l10n.defaults.time_picker,
 			date_picker  : l10n.defaults.date_picker,
@@ -306,13 +307,13 @@ window.CMB2 = (function(window, document, $, undefined){
 		return false;
 	};
 
-	$.fn.cleanRow = function( prevNum, group ) {
-		var $self = $(this);
-		var $inputs = $self.find( 'input:not([type="button"]), select, textarea, label' );
-		var $other  = $self.find('[id]').not( 'input:not([type="button"]), select, textarea, label' );
+	cmb.cleanRow = function( $row, prevNum, group ) {
+
+		var $inputs = $row.find( 'input:not([type="button"]), select, textarea, label' );
+		var $other  = $row.find('[id]').not( 'input:not([type="button"]), select, textarea, label' );
 		if ( group ) {
 			// Remove extra ajaxed rows
-			$self.find('.cmb-repeat-table .cmb-repeat-row:not(:first-child)').remove();
+			$row.find('.cmb-repeat-table .cmb-repeat-row:not(:first-child)').remove();
 
 			// Update all elements w/ an ID
 			if ( $other.length ) {
@@ -320,7 +321,7 @@ window.CMB2 = (function(window, document, $, undefined){
 					var $_this = $( this );
 					var oldID = $_this.attr( 'id' );
 					var newID = oldID.replace( '_'+ prevNum, '_'+ cmb.idNumber );
-					var $buttons = $self.find('[data-selector="'+ oldID +'"]');
+					var $buttons = $row.find('[data-selector="'+ oldID +'"]');
 					$_this.attr( 'id', newID );
 
 					// Replace data-selector vars
@@ -335,8 +336,8 @@ window.CMB2 = (function(window, document, $, undefined){
 		$inputs.filter(':checked').prop( 'checked', false );
 		$inputs.filter(':selected').prop( 'selected', false );
 
-		if ( $self.find('h3.cmb-group-title').length ) {
-			$self.find( 'h3.cmb-group-title' ).text( $self.data( 'title' ).replace( '{#}', ( cmb.idNumber + 1 ) ) );
+		if ( $row.find('h3.cmb-group-title').length ) {
+			$row.find( 'h3.cmb-group-title' ).text( $row.data( 'title' ).replace( '{#}', ( cmb.idNumber + 1 ) ) );
 		}
 
 		$inputs.each( function(){
@@ -385,11 +386,11 @@ window.CMB2 = (function(window, document, $, undefined){
 			}
 		});
 
-		return this;
+		return cmb;
 	};
 
-	$.fn.newRowHousekeeping = function() {
-		var $row         = $(this);
+	cmb.newRowHousekeeping = function( $row ) {
+
 		var $colorPicker = $row.find( '.wp-picker-container' );
 		var $list        = $row.find( '.cmb2-media-status' );
 
@@ -406,21 +407,10 @@ window.CMB2 = (function(window, document, $, undefined){
 			$list.empty();
 		}
 
-		return this;
+		return cmb;
 	};
 
-	cmb.afterRowInsert = function( $row, group ) {
-
-		var $focus = $row.find('input:not([type="button"]), textarea, select').first();
-		if ( $focus.length ) {
-			if ( group ) {
-				$('html, body').animate({
-					scrollTop: Math.round( $focus.offset().top - 150 )
-				}, 1000);
-			}
-			$focus.focus();
-		}
-
+	cmb.afterRowInsert = function( $row ) {
 		var _prop;
 
 		// Need to re-init wp_editor instances
@@ -499,12 +489,13 @@ window.CMB2 = (function(window, document, $, undefined){
 		cmb.idNumber = prevNum + 1;
 		var $row     = $oldRow.clone();
 
-		$row.data( 'title', $self.data( 'grouptitle' ) ).newRowHousekeeping().cleanRow( prevNum, true ).find( '.cmb-add-row-button' ).prop( 'disabled', false );
+		cmb.newRowHousekeeping( $row.data( 'title', $self.data( 'grouptitle' ) ) ).cleanRow( $row, prevNum, true );
+		$row.find( '.cmb-add-row-button' ).prop( 'disabled', false );
 
 		var $newRow = $( '<div class="postbox cmb-row cmb-repeatable-grouping" data-iterator="'+ cmb.idNumber +'">'+ $row.html() +'</div>' );
 		$oldRow.after( $newRow );
 
-		cmb.afterRowInsert( $newRow, true );
+		cmb.afterRowInsert( $newRow );
 
 		if ( $table.find('.cmb-repeatable-grouping').length <= 1 ) {
 			$table.find('.cmb-remove-group-row').prop( 'disabled', true );
@@ -525,7 +516,7 @@ window.CMB2 = (function(window, document, $, undefined){
 		cmb.idNumber      = prevNum + 1;
 		var $row          = $emptyrow.clone();
 
-		$row.newRowHousekeeping().cleanRow( prevNum );
+		cmb.newRowHousekeeping( $row ).cleanRow( $row, prevNum );
 
 		$emptyrow.removeClass('empty-row hidden').addClass('cmb-repeat-row');
 		$emptyrow.after( $row );
@@ -679,7 +670,8 @@ window.CMB2 = (function(window, document, $, undefined){
 			return;
 		}
 
-		$selector.timePicker( cmb.defaults.time_picker );
+		$selector.timepicker( 'destroy' );
+		$selector.timepicker( cmb.defaults.time_picker );
 	};
 
 	cmb.initDatePickers = function( $selector ) {
@@ -687,7 +679,7 @@ window.CMB2 = (function(window, document, $, undefined){
 			return;
 		}
 
-		$selector.datepicker( "destroy" );
+		$selector.datepicker( 'destroy' );
 		$selector.datepicker( cmb.defaults.date_picker );
 	};
 
@@ -716,7 +708,7 @@ window.CMB2 = (function(window, document, $, undefined){
 	cmb.makeListSortable = function() {
 		var $filelist = cmb.metabox().find( '.cmb2-media-status.cmb-attach-list' );
 		if ( $filelist.length ) {
-			$filelist.sortable({ cursor: "move" }).disableSelection();
+			$filelist.sortable({ cursor: 'move' }).disableSelection();
 		}
 	};
 
@@ -738,7 +730,7 @@ window.CMB2 = (function(window, document, $, undefined){
 				// Only Ajax on normal keystrokes
 				if ( betw( 48, 90 ) || betw( 96, 111 ) || betw( 8, 9 ) || evt.which === 187 || evt.which === 190 ) {
 					// fire our ajax function
-					cmb.doAjax( $self, evt);
+					cmb.doAjax( $self, evt );
 				}
 			},
 			paste : function() {
@@ -825,7 +817,7 @@ window.CMB2 = (function(window, document, $, undefined){
 	};
 
 	// function for running our ajax
-	cmb.doAjax = function($obj) {
+	cmb.doAjax = function( $obj ) {
 		// get typed value
 		var oembed_url = $obj.val();
 		// only proceed if the field contains more than 6 characters
@@ -833,55 +825,49 @@ window.CMB2 = (function(window, document, $, undefined){
 			return;
 		}
 
-		// only proceed if the user has pasted, pressed a number, letter, or whitelisted characters
+		// get field id
+		var field_id         = $obj.attr('id');
+		var $context         = $obj.closest( '.cmb-td' );
+		var $embed_container = $context.find( '.embed-status' );
+		var $embed_wrap      = $context.find( '.embed_wrap' );
+		var $child_el        = $embed_container.find( ':first-child' );
+		var oembed_width     = $embed_container.length && $child_el.length ? $child_el.width() : $obj.width();
 
-			// get field id
-			var field_id = $obj.attr('id');
-			// get our inputs $context for pinpointing
-			var $context = $obj.parents('.cmb-repeat-table  .cmb-row .cmb-td');
-			$context = $context.length ? $context : $obj.parents('.cmb2-metabox .cmb-row .cmb-td');
+		cmb.log( 'oembed_url', oembed_url, field_id );
 
-			var embed_container = $('.embed-status', $context);
-			var oembed_width = $obj.width();
-			var child_el = $(':first-child', embed_container);
-
-			// http://www.youtube.com/watch?v=dGG7aru2S6U
-			cmb.log( 'oembed_url', oembed_url, field_id );
-			oembed_width = ( embed_container.length && child_el.length ) ? child_el.width() : $obj.width();
-
-			// show our spinner
-			cmb.spinner( $context );
-			// clear out previous results
-			$('.embed_wrap', $context).html('');
-			// and run our ajax function
-			setTimeout( function() {
-				// if they haven't typed in 500 ms
-				if ( $('.cmb2-oembed:focus').val() !== oembed_url ) {
-					return;
+		// show our spinner
+		cmb.spinner( $context );
+		// clear out previous results
+		$embed_wrap.html('');
+		// and run our ajax function
+		setTimeout( function() {
+			// if they haven't typed in 500 ms
+			if ( $( '.cmb2-oembed:focus' ).val() !== oembed_url ) {
+				return;
+			}
+			$.ajax({
+				type : 'post',
+				dataType : 'json',
+				url : l10n.ajaxurl,
+				data : {
+					'action'          : 'cmb2_oembed_handler',
+					'oembed_url'      : oembed_url,
+					'oembed_width'    : oembed_width > 300 ? oembed_width : 300,
+					'field_id'        : field_id,
+					'object_id'       : $obj.data( 'objectid' ),
+					'object_type'     : $obj.data( 'objecttype' ),
+					'cmb2_ajax_nonce' : l10n.ajax_nonce
+				},
+				success: function(response) {
+					cmb.log( response );
+					// hide our spinner
+					cmb.spinner( $context, true );
+					// and populate our results from ajax response
+					$embed_wrap.html( response.data );
 				}
-				$.ajax({
-					type : 'post',
-					dataType : 'json',
-					url : l10n.ajaxurl,
-					data : {
-						'action': 'cmb2_oembed_handler',
-						'oembed_url': oembed_url,
-						'oembed_width': oembed_width > 300 ? oembed_width : 300,
-						'field_id': field_id,
-						'object_id': $obj.data('objectid'),
-						'object_type': $obj.data('objecttype'),
-						'cmb2_ajax_nonce': l10n.ajax_nonce
-					},
-					success: function(response) {
-						cmb.log( response );
-						// hide our spinner
-						cmb.spinner( $context, true );
-						// and populate our results from ajax response
-						$('.embed_wrap', $context).html(response.data);
-					}
-				});
+			});
 
-			}, 500);
+		}, 500);
 	};
 
 	$(document).ready(cmb.init);